+++ HISTORY.rst
@@ -3,7 +3,7 @@
 History
 -------
 
-1.1.1 (2019-??-??)
+1.1.1 (2019-08-16)
 ______________________
 * Fix minimum order number and warning the user, it also show in the price break.
 * Improve some spreadsheet formulas.